GroupedFileUploadToFile2: support files >2GB
Previously, we'd just crash. It's amazing we got this far until
experiencing this.

Someday soon, we _must_ move file upload into S3. At that point, this
change will be moot.

+++ common/src/main/scala/com/overviewdocs/models/BlobStorageRef.scala
@@ -4,5 +4,5 @@ package com.overviewdocs.models
   */
 case class BlobStorageRef(
   location: String,
-  nBytes: Int
+  nBytes: Long
 )
